api/requirements.txt
pyyaml
asyncio-nats-client
requests
flask
flask-restplus
matplotlib
mpld3
